move preprocessing to builder
