+++ builder/testdata/preprocessor/alphanumeric-external.yaml
