+++ builder/testdata/preprocessor/invalid-external.yaml
