+++ builder/testdata/preprocessor/preprocessing-stress.yaml
