+++ builder/testdata/preprocessor/valid-alias.yaml
