+++ builder/testdata/preprocessor/valid-external.yaml
