+++ builder/testdata/preprocessor/valid-remote.yaml
