+++ builder/util_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package graph
+package builder
 
 import (
 	"testing"
