+++ graph/task.go
@@ -87,9 +87,6 @@ type TaskOptions struct {
 
 	// DoPreprocessing is a property to decide whether we use Alias
 	DoPreprocessing bool
-
-	// GlobalAliases keeps track of all the Task native global aliases
-	GlobalAliases []byte
 }
 
 // UnmarshalTaskFromString unmarshals a Task from a raw string.
